requirements.txt
django==5.1.4
djangorestframework==3.15.2
sqlparse==0.5.3
asgiref==3.8.1
markdown==3.7
django-filter==24.3
pyjwt==2.10.1
djangorestframework-simplejwt==5.4.0
Pillow==11.1.0